[artf5798] : develop tests to fully test command threading system
Restored a test of repository-command concurrency. Commands for 3 repos are issued with a thread pool size of 2, so 2 should run concurrently and the third only after one of the first two finished. I found that the order of execution was random, so i couldn't assert that command1 ran concurrently with command2 and sequential with command3, but I set up a broad expectation of at least one sequential pair and one parallel pair.

The stateTransitionLogging was sometimes skipping events, so the test was failing intermittently -- I added synchronization which seems to have solved. Will not affect performance when state transitions are not logged (dev, prod).

* src/groovy/../AbstractCommand.groovy (logStateTransition)
added synchronization around the logging code, with a common lock shared by all commands

* test/integration/../CommandThreadingIntegrationTests.groovy (testBasicBlockingAndConcurrency)
restored test for repo-command blocking and concurrency


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2219 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f