[artf5798] : develop tests to fully test command threading system
Yet another fix for failing test on jenkins -- needed to modify "getExecutionOrder(cmd1, cmd2)" to be agnostic about which command ran first. Before this change, it assumed cmd1 was first, so the expectations were not always met if cmd2 ran first.

* test/integration/../CommandThreadingIntegrationTests.groovy (getExecutionOrder)
now just looks for overlaps in the runtimes, irrespective of which began first. before, assumed that "cmd1" was the one to run first.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2221 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f