[artf6220] : Replica server requires site admin privs for too much
Partial implementation of the artifact. User credentials that match those of the current CTF user configured for the replica will be granted ROLE_ADMIN on the replica (provided the credentials also authenticate against CTF).

* grails-app/services/../CtfRemoteClientService.groovy (authenticateUser, getUserDetails, getGrantedAuthorities)
authenticateUser() now checks the credentials against the CtfServer table. If matching, a boolean flag is passed through to getGrantedAuthorities(isLocalAdmin), which will provide ROLE_ADMIN for the user.
 


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2342 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f