[artf6220] : Replica server requires site admin privs for too much
Added a check to see if the user session can access IScmAppSoap.getScmAdapterNames(). If yes, we will grant local ROLE_ADMIN to the auth session. This will currently fail since the web method requires SuperUser permission, but I am opening issue in CTF to request the permissions adjustment.

Also, a small change to CtfRemoteStrategy to allow compiling (subclasses did not match signature exactly) -- worked okay in Grails, but not in IDE compile. now works in both.

* src/groovy/../CtfRemoteStrategy.groovy (login)
added exceptions thrown in subclass to abstract signature

* grails-app/services/../CtfRemoteClientService.groovy (getUserDetails)
now checks for access to IScmAppSoap.getScmAdapterNames() if the user session is not already granted ROLE_ADMIN.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2359 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f