[artf6297] : create repo list that shows backup jobs
Added the list of repos with associated backup job info to the Backup Schedule screen. This uses the same handler logic as the individual repo backup scheduler, which was modified to handle both submissions.

* grails-app/views/repo/_backupScheduleForm.gsp
Added the list of repositoryBackupJobs to the scheduler form when present in the page view model. This had to go here so that all the form elements could belong to the same form.

* grails-app/i18n/messages.properties
new keys for Backup Schedule view

* src/groovy/../DumpBean.groovy
added "cloud" (Boolean) property for distinguishing cloud backups

* grails-app/controllers/../RepoController.groovy (updateBkupSchedule, bkupScheduleMultiple, formatSchedule)
changed updateBkupSchedule to handle submission of individual and multiple repo schedules; added view preparation to bkupScheduleMultiple (uses new helper formatSchedule)



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2404 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f