[artf6328] : Convert to Teamforge mode fails unexpectedly
RemoteMasterException could not be constructed with "null" second argument due to signature ambiguity. This fix removes one of the two-argument constructors and adjusts the one client usage.

* grails-app/services/../CtfRemoteClientService.groovy (addExternalSystemReplica)
adjusted usage of RemoteMasterException so that one of the two-arg constructors could be removed

* src/groovy/../RemoteMasterException (constructor)
removed the least-used two-arg constructor, as three-arg constructor could be easily used instead for this case

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2418 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f