[artf6304] : communicate with cloud API
Initial, sunny-day implementation of registering organization and administrator. Not fully baked -- need more info on how to report error conditions. 

* grails-app/congroller/../SetupCloudServicesController.groovy (createAccount)
added the call to the remote service for creating an account

* grails-app/servies/../CloudServiesRemoteClientService.groovy (createAccount)
New method makes rest call to create organization and, on success, records the credentials in our database

* grails-app/i18n/messages.properties
general error message for failure to create the codesion account

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2434 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f