[artf6412] : Backup Settings UI - change backup settings to match selected item
Multiple-repo backup scheduling now sets the form state from the first item clicked which has a scheduled job. 

* controllers/RepoController.groovy (bkupScheduleMultiple, getListViewSelectedIds, updateBkupSchedule, dumpOptions, downloadDumpFile, deleteDumpFiles, deleteMultiple)
bkupScheduleMultiple now adds schedule info each job that will be listed, so the data can be used to set the form; the other methods were updated to use ControllerUtil.getListViewSelectedIds() instead of the local private method, which was deleted

* taglib/CsvnTagLib.groovy (listViewSelectItem)
now copies an "onClick" attribute from the custom tag to the resulting input tag

* views/repo/_backupScheduleForm.gsp
added javascript for copying the schedule settings from a row in the repository list to the form. This behavior can happen exactly once per page refresh, and is disabled when the user manipulates the form directly.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2489 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f