[artf6467] : don't del update folder when errors occur
I added error detection to the copyfiles section of the install-updates script. If xcopy returns with an error code > 1, the error is indicated in the log and the updates folder is retained (error code 1 means no files to copy). I had to remove the "/C" option from xcopy (continue on error), which means it now stops as soon as there is a file lock or sharing violation. It used to continue and return the success code even if some of the files could not be replaced.

* install-updates.bat
added error logging with xcopy exit codes of 2 or greater. If this is the case, the /updates folder is retained and the script can be run again. Xcopy now stops as soon as an error is encountered.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2510 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f