[artf6470] : add svn load job
made some changes to catch and log errors in loading a dump file

* grails-app/services/../CommandLineService.groovy (startProcessWithInputStream)
now destroys proc if the input stream is interrupted -- this should be visible to caller as non-zero exit code on the proc

* src/groovy/../RepoLoadException.groovy
new exception class to indicate dump file load problems

* grails-app/jobs/../RepoLoadJob.groovy
now catches and logs exceptions

* grails-app/services/../SvnRepoService.groovy (loadDumpFile)
now throws RepoLoadException when there is an issue loading the dump file (non-zero exit code from external proc). The dumpFile and output capture file are retained in this case.




git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2533 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f