[artf4998]: Load a repository
modified loadDumpFile() to retry deletion of dumpfile up to 5 times after successful load. This is chiefly needed because of the regular failure of a unit test on Win64, where the dumpfile will not delete.

* services/../SvnRepoService.groovy (loadDumpFile, deleteWithRetry)
loadDump file now calls deletewithretry(file) instead of file.delete() to remove the load file and log.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2562 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f