[artf6550] : Functional test failure: error migrating 2->2.1 in db init
Actually -- i turns out the "restart link" code was the cause of all the functional test errors (added in cmmt7070 for artf6230). In most cases, the link is not rendered in the page, so the "observe" method is applied to a null element. In the real world, this shows up in the javascript console as an error but doesn't otherwise affect the page. 

With this commit now checks for null.

* views/status/index.gsp (window load observer)
now checks for existence of restartLink before applying observer



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2563 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f