[artf6438] : add ajax uniqueness checker to domain name field
refactored the domain and login name uniqueness checks to show the validation error if there is one or the availability.

* controllers/../SetupCloudServicesController.groovy (checkDomainAvailability, checkLoginAvailability)
now sends the result message directly to the ajax client  -- this lets us display any validation error without preparing the messages in javascript. A second "tokenStatus" property was added to the response JSON, for signalling ok/not ok.
 
 * views/../signup.gsp
 now constructs simplified CloudTokenAvailabilityChecker, since most messages can be sent from server
 
 * web-app/js/application.js (CloudTokenAvailabiltyChecker)
 now displays result messages from server. uses the "tokenStatus" field in the response to indicate which icon (ok or warning) to show with the message.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2572 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f