[artf6555] : debugging improvements for production images
A couple of changes to this: removed quotes on debug params, as they break java startup on linux; added logic to the DistDebug script to support windows build.

* scripts/DistDebug.groovy 
added logic for windows dist (operate on /updates/ folder)

* csvn-service-wrapper/(linux|solaris|windows)/conf/csvn-wrapper.conf
removed quotes on debug params. Probably the windows folder should be deleted since it's not in use, but I updated here anyway.

* svcwrapper/conf/wrapper.conf
removed quotes on debug params



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2579 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f