[artf6622] : use a backup when creating repo
This commit adds a rough-cut UI for selecting a backup file to initialize a new Repo. The feature is not functional yet, but all the dump files from extant repos are now presented for selection in a tree view, and exactly one dump file can be selected by single clicking.  This uses the the "Dynamic Drive Simple Tree Menu" library, which is free for all uses.

Template selection has been left as-is for the time being, though the "create tags/branches/trunk" option has been moved into a div container like that in Mark's mock up, which is swapped out when Backup is selected as the init option.

* grails-app/views/repo/create.gsp
included DD simpletreemenu.js and css; added JS for showing / hiding the backups and templates selection divs; added JS for fetching and displaying tree of backup files (ajax).

* grails-app/i18n/messages.properties
new keys for "Backup" and "Initialize" labels

* grails-app/controllers/../RepoController.groovy (dumpFileListAll)
new action to render all the dump files as JSON

* web-app/css/simpletree.css
the css for the treeview

* web-app/js/simpletree.js
the js for the treeview



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2593 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f