[artf6682] : Jobs UI sometimes doesn't show job executions
Sometimes when you click on the jobs page, the "finished jobs" section doesn't show jobs that have run. The problem is that the supporting service (JobsInfoService) is not created until you view the page, so it is therefore not listening to quartz events prior to the first page view.

Fix is to inject the service at bootstrap, so that the instance is created and registered with the quartz scheduler.

* grails-app/conf/BootStrap.groovy
added injection reference for jobsInfoService, so tha the instance is created at startup.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2597 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f