[artf6622] : use a backup when creating repo
modified unzip code to use commandline "unzip" on unix, to retain repo permissions; refactored FileUtil as a spring bean and moved the archive methods from SvnRepoService to this.

* grails-app/services/../SvnRepoService.groovy (createHotCopy, loadHotCopy, archiveDirectory, recursiveArchiveDirectory)
now uses spring bean "fileUtil" for zip and unzip operations; moved the archiveDirectory and helper recursiveArchiveDirectory to this helper

* grails-app/conf/spring/resources.groovy
new bean, "fileUtil," is named

* src/groovy/../FileUtil.groovy (archiveDirectory, recursiveArchiveDirectory, extractArchive, extractArchiveCommonsCompress)
refactored to use spring dependencies for OS and CommandLine services; moved create archive methods from SvnRepoService to here; renamed unzipFileIntoDirectory() to extractArchive(), which now uses Unix "unzip" command line call on that platform; extractArchiveCommonsCompress() is left here for reference -- it extracts via the Apache commons lib.

* test/integration/../SvnRepoServiceIntegrationTests.groovy (testLoadHotcopyAsBackup)
this test now sets the execute bit for one of the hook script templates, and verifies that the bit is set in the resulting target repo (unix only)

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2620 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f