[artf6614] : Problems capturing network stats on Windows
It appears we have always only captured stats for the first non-loopback interface, irrespective of the user selection in the server settings. This commit allows user changes in the server settings to update the NetworkService.selectedInterface, which is used by the stat collect job. The stored value is also consulted at bootstrap.

* grails-app/services/../NetworkingService.groovy (bootStrap, setSelectedInterface)
new setter for configuring the selected IF; bootstrap calls this if Server.netInterface has a value

* grails-app/controllers/../ServerController.groovy (update)
now calls networkingService.groovy.setSelectedInterface() on successful update

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2626 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f