[artf6797]: Configure JVM to use a proxy
Here are a few changes to the OCN controller and the Cloud Services client service which allow use of the proxy settings. I also updated the ProxyTest.groovy to accurately show success/failure (this script is best run in IDE so that all the dependent libs are available).

* test/other/ProxyTest.groovy
added exception handling and messaging for axisClient() and restClient() to show sucess when we get the expected result from the origin server

* controller/../OcnController.groovy
converted URL fetching to use HTTPBuilder, which supports proxy auth

* services/../CloudServicesRemoteClientService.groovy (createRestClient)
added code to configure client proxy settings if needed

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2674 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f