[artf6835]: "older ctf" errror message
When connecting to CTF during setup, certain error conditions throw a RemoteMasterException without a message key property -- and the controller was writing "older ctf" as the default message in the absence of a key. This changes the SetupReplicaController to use the RemoteMasterException.message property if the key property is null. Additionally, the CtfRemoteClientService is changed to provide the general error message when the cause is unknown (instead of "auth" error).

* services/../CtfRemoteClientService.groovy (login)
unknown errors now contain the general error instead of "auth"

* controllers/../SetupReplicaController.groovy (replicaSetup)
RemoteMasterException without message key will now default to the exception message instead of "older ctf".

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2686 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f