[artf6831]: unit testing
added a test to NetworkingServiceTests that verifies the VM proxy properties are set as expected. Also, made a small tearDown fix to the ServerConfServiceTests, to clear the dynamic methods added to CtfServer for mocking.

* test/unit/../NetworkingServiceTests.groovy (setUp, tearDown, testSetHttpProxy)
tests that the networkingService.setHttpProxy method produces the desired effects in the VM properties

* test/unit/../ServerConfServiceTests.groovy (tearDown)
now clears the metaClass changes that were applied to CtfServer in order to mock this domain class properly

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2705 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f