[artf6852] : Logging settings API
I noticed that Spring security was using status code 302 to redirect to /login/denied when our user credentials lacked the necessary role grant for an action, which is not appropriate for an API client. This change registers a custom "AccessDeniedHandler" for the REST security filter chain, so that for these urls a 401 error is sent instead.

* src/groovy/../RestAccessDeniedHandler.groovy
custom implementation of Spring Security AccessDeniedHandler, which sends 401 unauthorized message to clients instead of redirecting.

* test/functional/../LoggingApiTests.groovy (testLoggingGet)
changed an assert to expect 401 instead of 302 for user authorization trying to access admin resource

* grails-app/conf/spring/resources.groovy (basicExceptionTranslationFilter, restAccessDeniedHandler)
new RestAccessDeniedHandler is instantiated and injected into the basicExceptionTranslationFilter

* grails-app/messages.properties
new message: api.error.401

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2730 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f