[artf6898] : Cloud Services integration test failure
Codesion has a new cert issued on 12/22/11, from a new issuer. The new root cert is here, and has been imported (along with the codesion.com cert) to the client keystore we ship.

* svn-server/data/certs/cloud_services_root_ca.crt
new root CA cert

* svn-server/data/certs/svnedge-client.jks
updated the client keystore with the root CA and the codesion.com domain cert.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2740 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f