[artf6981] : API to upload/replace file
This API allows an arbitrary hook script to be placed in a repo's hooks directory. The file content is presumed to occupy the request body, and the name is derived from the url. Both text and binary executables can be uploaded if the content-type and encoding are correct.

* grails-app/conf/UrlMappings.groovy
api url pattern needed to capture additional path info after the entity id to params.cgiPathInfo 

* grails-app/controllers/../HookRestController.groovy (restUpdate)
web method to write the request body to the hooks folder for the given repo
 
 * grails-app/i18n/messages.properties
 new strings for api error messaging
 
 * src/groovy/../ControllerUtil.groovy (getFileFromRequest)
 new helper method to extract a file from the request using stream operations
 
 * test/functional/../ApiTestHelper.groovy (createRepo)
 helper method to create a test repo
 
 * test/functional/../HookApiTests.groovy 
 REST client tests for plain and binary file PUTs 


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2781 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f