[artf6981] : API to upload/replace file
Added support for POST of new hook file. Same as PUT, except an existing file by the same name will prevent it from succeeding.

* controller/api/HookRestController.groovy (restSave)
new POST handler that simply forwards to PUT, except a parameter is added to indicate that overwrites are not allowed.

* services/../SvnRepoService.groovy (createHook)
new method to create a hook script, but fails if an existing file is in the way

* test/functional/../HookApiTests (testHookPost)
new test that post to existing file location will fail, but new file will succeed.



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2818 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f