[artf6899]: functional test failures
Additional fix for TF conversion test failures

* AbstractConversionFunctionalTests.groovy (assertConversionSucceeded, assertConversionDidNotSuceeded)
These were testing for "Server version" on the status page, which apparently does not appear when the packagesUpdateService has problems bootstrapping. I removed the test, since this is not germane to TF conversion success / failure -- only the presence of the TF master url indicates success / failure.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2842 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f