[artf7042] : work on tables
Ongoing work. Adapted the repo-detail layout and switched to bootstrap for nav tabs and tables.

* grails-app/views/common/_tabs.gsp
switched this component to render bootstrap-style tabs
 
* grails-app/views/common/_fileList.gsp
switched the file list component to use bootstrap classes on tables

* grails-app/views/layouts/repoDetail.gsp
changed this layout to follow mark's mockup for the repo details and tabbed content, though left nav is still rendered



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2887 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f