[artf7062] : Fix javascript (particularly for list-view behaviors)
This restores much of the JS functionality for the repository list views and switches the modals to bootstrap style. Because bootstrap modal relies more on html than our previous lib, much of the funcionality was moved from listView.js to CsvnTagLib.listViewAction.

* views/common/_listViewResources.gsp (removed)
* views/repo/bkupScheduleMultiple.gsp
* views/repo/dumpFileList.gsp
* views/repo/hooksList.gsp
* views/repo/list.gsp
* views/repoTemplate/list.gsp
* views/setupCloudServices/selectUsers.gsp
After switching from prototype to jquery, _listViewResources.gsp no longer had anything in it except an include of the listView.js, so I deleted the gsp fragment entirely, and replaced usages with a direct call to include the JS file.

* views/layouts/main.gsp
moved the 'jquery' include from the foot to the head of the file -- needed to have jquery available ahead of the content panel rendering.

* web-app/js/bootstrap-modal.js
added this plugin, but not sure if it's needed
 
 * web-app/js/listView.js
 adapted from prototype to jquery syntax; removed observers for  listViewAction buttons -- now written into html by the custom tag 
 
 * grails-app/taglib/CsvnTagLib.groovy (listViewActionButton)
 adapted to use bootstrap / jquery internally, but invocation is the same from GSP layer
 
 * grails-app/i18n/messages.properties
 added "default.confirmation.title"
 






git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2905 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f