[artf7063] : Convert tables to jquery Datatables
Converted hooks and dump file listing to datatable
 
 * grails-app/views/common/_fileList.gsp
 * grails-app/views/repo/dumpFileList.gsp
 * grails-app/views/repo/hooksList.gsp
updated file-list component to render datatable. small change was needed to dumps and hooks list page (replacing listview js include at head with link to the css file)

* src/groovy/../ControllerUtil.groovy (decorateFileClass)
* controllers/../LogController.groovy (list)
* controllers/../RepoController.groovy (dumpFileList, hooksList)
created ControllerUtil helper to add "size" and "date" bean props to the File class, so that it is more jstl-friendly. LogController and RepoController are using. 



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2980 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f