[artf7029]: Update look and feel of viewvc templates
The About dialog in ViewVC now obtains the release version from a local text file created in the Dist build script, so that it is always the same as the console version.

* viewvc/docroot/include/header.ezt
* viewvc/docroot/app_version.txt
header now includes app_version.txt in the about dialog for the version info. This file is clobbered by the Dist build target with the actual version number.

* scripts/Dist.groovy (rearrangingArtifacts)
now echos the grails app.version property (from application.properties) to www/viewvc/app_version.txt during the Dist build. 







git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3044 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f