[artf7029]: Update look and feel of viewvc templates
Fixed faulty logic in Windows for writing the "viewvc-integrated" template path to the viewvc.conf file.

* src/groovy/../ConfigUtil.groovy (viewvcTemplateDirIntegrated)
This method would return the default viewvc path if it was configured in Config.groovy (which it is for Windows). Changed it consult a new key instead, "svnedge.svn.viewvcTemplatesIntegratedPath", which is currently not defined for any platform.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3046 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f