[artf7176] : Improve Discover Repositories performance
Fix for test failure -- the opening assertion that there be 0 repos is failing on some platforms, but isn't really germane to the test anyway. So refactored to test changes in repo count against status quo ante instead of 0. 

* test/integration/../SvnRepoServiceIntegrationTests.groovy (testSyncRepositoriesPerformance)
removed expectation that no repos be present at beginning of the test.

 
 


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3067 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f