[artf7177] : Add new repository job to Verify repository
A couple of follow up fixes. 1) when job is created, message says only errors will be reported and, if mail settings are not configured, directs users there. 2) small change to maintain the form state after creating a verify job.

* i18n/messages.properties
new / updated messages

* views/repo/_backupScheduleForm.gsp
updated logic to select "verify" if previously submitted

* controller/../RepoController.groovy (bkupSchedule)
if there's no existing DumpBean, the form was defaulting to job type "none", but now we might have created a verify job, so the type setting needs to be sticky.

 



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3086 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f