[artf7080] : Hotcopy backup fails when there are symlinks
This commit adds a command-line zip utility method for creating hotcopies, which is used as a fall-back if the java-based zip fails due to symlinks. In this case, if "zip" is not available on the path, then we cannot create the archive and the hotcopy will fail.

* src/groovy/../FileUtil.groovy (archiveDirectory, archiveDirectoryUsingExternalTool, extractArchive)
method archiveDirectory will catch "StringIndexOutOfBoundsException" during zip operation and attempt to zip via alternate archiveDirectoryUsingExternalTool(). extractArchive() needed change to the chmod command it executes, as the commandline zip util results in files with 6-digit unix mode info.

* CommandLineService.groovy (startProcess)
added support for setting a current working directory for the process via the environment map entry "PWD". In this case, it is used to create a zip archive from one directory above.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3100 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f