[artf7456] : Be more resilient about SIGAR errors
Broadening the type of exceptions caught by the service for checking process ids. If an exception is caused by the failure to load the underlying SIGAR lib, we now assume the process is alive and well if a pid file is present, as it cannot be checked.

* grails-app/servies/../OperatingSystemService.groovy (getProcessExists)
catches all exceptions now, mostly returning false when an exception occurs.

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3223 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f