[artf7222] Improve probability of all tests passing
* test/integration/com/collabnet/svnedge/TestJobHelper.groovy

Occasionally Quartz does not run the job under test until after the test
times out waiting for it.  Since in normal cases, the wait time is less than
a second, reducing the wait time to 30s.  Adding a thread dump before and
after the wait to see if anything valuable can be gleaned. 



git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3277 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f