[artf7222] Improve probability of all tests passing
Giving up testing using quartz to trigger the job as it requires debugging of
quartz on a system which exhibits the problem.  And it would be preferable 
to see if the bad behavior still exists in a more recent version of quartz
before further debugging of 1.7.3.  Just going to invoke the job inline for now.

* test/integration/com/collabnet/svnedge/statistics/job/DeleteStatJobIntegrationTests.groovy
* test/integration/com/collabnet/svnedge/statistics/job/ConsolidateStatJobIntegrationTests.groovy
* test/integration/com/collabnet/svnedge/statistics/job/StatCollectJobIntegrationTests.groovy
* test/integration/com/collabnet/svnedge/statistics/job/StatCountJobIntegrationTests.groovy
  Need to manually set the dependencies, since quartz plugin is not being used

* test/integration/com/collabnet/svnedge/TestJobHelper.groovy
  (executeJob): Directly invoking job.execute instead of having a trigger fire.




git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3282 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f