[artf7741] Every selection [ON | OFF] of ‘Apache Encryption’ produces a separate link for ‘Advanced Configuration’. [artf7714] Upgrade to Bootstrap 2.3.0
The generally preferred method for reading the state of checkboxes is with 'prop()', setting
the 'attr' works and will affect a read via 'prop', but a user clicking the checkbox will 
not affect 'attr()'.
The change to the attr() method to reflect the html and not user action was a change jquery 
tried to make in 1.6, but got too much pushback, so reverted, but finally made the change in 1.9.

Since we were using 'attr' as a synonym for 'prop' in all cases I found, changing them all 
to 'prop'.

* grails-app/views/repoTemplate/list.gsp
* grails-app/views/server/editMail.gsp
* grails-app/views/server/editMonitoring.gsp
* grails-app/views/server/edit.gsp
* grails-app/views/server/editAuthentication.gsp
* grails-app/views/gettingStartedWizard/_serverSettings.gsp
* grails-app/views/repo/_backupScheduleForm.gsp
* grails-app/views/repo/list.gsp
* grails-app/views/repo/dumpOptions.gsp
* grails-app/views/repo/loadOptions.gsp
* grails-app/views/repo/create.gsp
* grails-app/views/user/create.gsp
* grails-app/views/user/edit.gsp
* grails-app/views/status/index.gsp
* grails-app/views/setupTeamForge/ctfUsers.gsp
* grails-app/views/setupCloudServices/createUserLogins.gsp
* grails-app/views/setupCloudServices/signup.gsp


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3452 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f