[artf7985] Database error logged when setting up replication
Use of afterLoad method in AdvancedConfiguration is triggering version to 
increment. An alternate solution to the problem would introduce transient 
properties for the fields which have defaults for null persistent values.
The chosen solution is more localized. The code was not using the version to
avoid dirty writes of the form:
User A visits advConfig page
User B visits advConfig page
User B updates (saves) particular properties on the page
User A saves potentially conflicting property changes.

While it is possible to use the version property in this scenario, we were
not, so User A's changes will win (last write wins scenario). So the version
property is helping with handling the long transaction conflict, but is only causing
problems due to concurrent thread's thinking each of their AdvancedConfiguration
objects were out-of-date. There is almost no chance of a "last write wins" 
being a problem for AdvancedConfiguration on a concurrent thread level, 
so disabling optimistic locking is the simpler solution. 

* grails-app/domain/com/collabnet/svnedge/domain/AdvancedConfiguration.groovy
  Don't map the version column. Disables optimistic locking.

* src/java/com/collabnet/svnedge/schema/MS021DropVersionFromAdvConfig.java
* src/java/com/collabnet/svnedge/schema/DatabaseInitializer.java
  New migration script to drop ADVANCED_CONFIGURATION.VERSION column


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3662 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f